Also consider moving clip's markers for snapping. Fixes #661

+++ src/timeline2/model/clipmodel.cpp
@@ -91,7 +91,7 @@ int ClipModel::construct(const std::shared_ptr<TimelineModel> &parent, const QSt
     return id;
 }
 
-void ClipModel::allSnaps(std::vector<size_t> &snaps)
+void ClipModel::allSnaps(std::vector<int> &snaps)
 {
     m_clipMarkerModel->allSnaps(snaps);
 }
